modularity improvements for switch to ui and to not get lost in adding additional budget components
