+++ Budget/WorkbookResult.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Budget
+{
+    internal class WorkbookResult
+    {
+        public bool Success { get; set; }
+        public string Message { get; set; }
+    }
+}
